am62x: PINMUX: Misra C issue Fix MISRA.ETYPE.CATEGORY.DIFFERENT.2012
-The operands of an operator in which the usual
 arithmetic conversions are performed do not have
 the same essential type category.

Fixes: SITSW-5882

+++ source/drivers/pinmux/am62x/pinmux.c
@@ -171,13 +171,13 @@ void Pinmux_unlockMMR(uint32_t domainId)
     {
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_PADCFG_CTRL0_CFG0_BASE);
         /* Lock 0 */
-        kickAddr = (volatile uint32_t *) (baseAddr + CSL_MAIN_PADCONFIG_LOCK0_KICK0_OFFSET);
+        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)CSL_MAIN_PADCONFIG_LOCK0_KICK0_OFFSET);
         CSL_REG32_WR(kickAddr, KICK0_UNLOCK_VAL);   /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK1_UNLOCK_VAL);   /* KICK 1 */
 
         /* Lock 1 */
-        kickAddr = (volatile uint32_t *) (baseAddr + CSL_MAIN_PADCONFIG_LOCK1_KICK0_OFFSET);
+        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)CSL_MAIN_PADCONFIG_LOCK1_KICK0_OFFSET);
         CSL_REG32_WR(kickAddr, KICK0_UNLOCK_VAL);   /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK1_UNLOCK_VAL);   /* KICK 1 */
@@ -187,13 +187,13 @@ void Pinmux_unlockMMR(uint32_t domainId)
     {
         baseAddr = (uint32_t) AddrTranslateP_getLocalAddr(CSL_MCU_PADCFG_CTRL0_CFG0_BASE);
         /* Lock 0 */
-        kickAddr = (volatile uint32_t *) (baseAddr + CSL_MCU_PADCONFIG_LOCK0_KICK0_OFFSET);
+        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)CSL_MCU_PADCONFIG_LOCK0_KICK0_OFFSET);
         CSL_REG32_WR(kickAddr, KICK0_UNLOCK_VAL);   /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK1_UNLOCK_VAL);   /* KICK 1 */
 
         /* Lock 1 */
-        kickAddr = (volatile uint32_t *) (baseAddr + CSL_MCU_PADCONFIG_LOCK1_KICK0_OFFSET);
+        kickAddr = (volatile uint32_t *) (baseAddr + (uint32_t)CSL_MCU_PADCONFIG_LOCK1_KICK0_OFFSET);
         CSL_REG32_WR(kickAddr, KICK0_UNLOCK_VAL);   /* KICK 0 */
         kickAddr++;
         CSL_REG32_WR(kickAddr, KICK1_UNLOCK_VAL);   /* KICK 1 */
